+++ src/mix_hex_http_httpc.erl
@@ -1,4 +1,4 @@
-%% Vendored from hex_core v0.18.0 (0f47da2), do not edit manually
+%% Vendored from hex_core v0.18.0 (d6a6a5a), do not edit manually
 
 %% @doc
 %% httpc-based implementation of {@link mix_hex_http} contract.